app: using uint32 as boxID

+++ app/config.go
@@ -8,7 +8,7 @@ import (
 type Config struct {
 	// Every instance belongs to a certain box.
 	// boxID: [1, 255)
-	BoxID int64 `toml:"box_id"`
+	BoxID uint32 `toml:"box_id"`
 	// Every instance has its own unique instanceID
 	InstanceID uint32 `toml:"instance-id"`
 	// Every instance belongs to a certain rack.
